requirements.txt
aiohttp==3.7.3
aiohttp-session==2.9.0
aiohttp-jinja2==1.4.2
aiomysql==0.0.21
bcrypt==3.2.0
cryptography==3.3.1
Jinja2==2.11.2
jinjasql==0.1.8
pydantic==1.7.3
pipenv==2020.11.15
python-dotenv==0.17.1